+++ tests/stubs/Queries/TestEnumResolvers.php
@@ -2,9 +2,7 @@
 
 namespace Butler\Graphql\Tests\Queries;
 
-if (PHP_VERSION_ID >= 80100) {
-    require_once __DIR__ . '/../enums.php';
-}
+require_once __DIR__ . '/../enums.php';
 
 class TestEnumResolvers
 {